+++ src/index.css
@@ -44,3 +44,29 @@
 .styleProgress::-webkit-progress-bar {
   background-color: #b0d8d8; /* Background behind the bar */
 }
+
+/* job Application */
+.application-two-lines {
+  display: -webkit-box;
+  -webkit-line-clamp: 3;
+  -webkit-box-orient: vertical;
+  overflow: hidden;
+  text-overflow: ellipsis;
+}
+
+@media screen and (min-width: 1024px) {
+  .application-two-lines {
+    -webkit-line-clamp: 2;
+  }
+}
+
+/* Application Check Box */
+
+.custom-checkbox {
+  display: none; /* Hide the native checkbox */
+}
+
+.appCateCheckBox {
+  z-index: 0;
+  transform: scale(1.2); /* Increase the size by 1.5 times */
+}
